update App.js, move information to PostPage.js
