+++ instagram-app/src/components/CommentSection/CommentSection.css
@@ -1,0 +1,18 @@
+/* CommentSection CSS */
+.comments {
+  text-align: left;
+  display: flex;
+  margin-left: 22%;
+  line-height: 0.1;
+}
+
+.newComment {
+  text-align: left;
+  margin-left: 22%;
+  margin-top: -10px;
+}
+
+.input {
+  margin-left: -45%;
+  margin-bottom: 50px;
+}
