+++ instagram-app/src/components/PostInput.js
@@ -7,7 +7,7 @@ const PostInput = props => {
     <form>
       <input type="text" value={props.comment} onChange={props.changeHandler} placeholder="Add comment" name="comment"/>
     </form>
-    <button onClick={props.change}>Click</button>
+    <button onClick={props.change}>Post</button>
   </div>
 
   );
